+++ libs/community/tests/unit_tests/embeddings/test_ovhcloud.py
@@ -23,10 +23,3 @@ def test_ovhcloud_empty_access_token_should_not_raise_error() -> None:
         model_name="multilingual-e5-base", region="kepler", access_token=""
     )
     assert isinstance(llm, OVHCloudEmbeddings)
-
-
-def test_ovhcloud_embed_documents() -> None:
-    llm = OVHCloudEmbeddings(model_name="multilingual-e5-base")
-    docs = ["Hello", "World"]
-    output = llm.embed_documents(docs)
-    assert len(output) == len(docs)
